Renomeacoes de disciplina para aula e prova para avaliacao
