+++ src/main/java/br/edu/ifnmg/poo/trabalhofinalpoo/gui/CadastrarAvaliacao.form
@@ -680,7 +680,7 @@
 
               <Layout class="org.netbeans.modules.form.compat2.layouts.support.JScrollPaneSupportLayout"/>
               <SubComponents>
-                <Component class="javax.swing.JList" name="lstDisciplinas">
+                <Component class="javax.swing.JList" name="lstAulas">
                   <Properties>
                     <Property name="background" type="java.awt.Color" editor="org.netbeans.beaninfo.editors.ColorEditor">
                       <Color blue="ff" green="ff" red="ff" type="rgb"/>
@@ -692,7 +692,7 @@
                       <Color blue="0" green="0" red="0" type="rgb"/>
                     </Property>
                     <Property name="model" type="javax.swing.ListModel" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
-                      <Connection code="lstDisciplinasModel" type="code"/>
+                      <Connection code="lstAulasModel" type="code"/>
                     </Property>
                   </Properties>
                   <AuxValues>
