+++ src/main/java/br/edu/ifnmg/poo/trabalhofinalpoo/gui/EditarAvaliacao.form
@@ -141,7 +141,7 @@
                   <Color blue="0" green="0" red="0" type="rgb"/>
                 </Property>
                 <Property name="model" type="javax.swing.ListModel" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
-                  <Connection code="lstDisciplinasModel" type="code"/>
+                  <Connection code="lstAulasModel" type="code"/>
                 </Property>
               </Properties>
               <AuxValues>
