+++ src/main/java/br/edu/ifnmg/poo/trabalhofinalpoo/gui/EditarAvaliacao.java
@@ -11,43 +11,46 @@ import br.edu.ifnmg.poo.trabalhofinalpoo.dao.AulaDao;
 import br.edu.ifnmg.poo.trabalhofinalpoo.entity.Avaliacao;
 import br.edu.ifnmg.poo.trabalhofinalpoo.entity.Discente;
 import br.edu.ifnmg.poo.trabalhofinalpoo.entity.Aula;
-import java.awt.event.MouseEvent;
 import java.util.List;
 import javax.swing.DefaultListModel;
 
 /**
  *
- * @author Fellipe
+ * @author felipe
  */
-public class CadastrarProva extends javax.swing.JFrame {
+public class EditarAvaliacao extends javax.swing.JDialog {
     
-    private DefaultListModel<Avaliacao> lstAvaliacaoModel;
+    private final CadastrarAvaliacao cadastroProva;
+    
+    private Avaliacao avaliacaoEmEdicao;
     
     private DefaultListModel<Discente> lstDiscentesModel;
     
-    private DefaultListModel<Aula> lstDisciplinasModel;
-    
-    private int indiceAvaliacaoSelecionada;
+    private DefaultListModel<Aula> lstAulasModel;
 
     /**
-     * Creates new form CadastrarProva
+     * Creates new form EditarAvaliacao
      */
-    public CadastrarProva() {
-        lstAvaliacaoModel = new DefaultListModel<>();
-        List<Avaliacao> avaliacoes = new AvaliacaoDao().localizarTodos();
-        lstAvaliacaoModel.addAll(avaliacoes);
+    public EditarAvaliacao(Avaliacao avaliacao, 
+        CadastrarAvaliacao cadastroProva, boolean modal) {
+        super(cadastroProva, modal);
         
-        // exibindo os alunos
         lstDiscentesModel = new DefaultListModel<>();
         List<Discente> discentes = new DiscenteDao().localizarTodos();
         lstDiscentesModel.addAll(discentes);
         
-        lstDisciplinasModel = new DefaultListModel<>();
-        List<Aula> disciplinas = new AulaDao().localizarTodos();
-        lstDisciplinasModel.addAll(disciplinas);
+        lstAulasModel = new DefaultListModel<>();
+        List<Aula> aulas = new AulaDao().localizarTodos();
+        lstAulasModel.addAll(aulas);
         
         initComponents();
+        
+        this.cadastroProva = cadastroProva;
+        this.avaliacaoEmEdicao = avaliacao;
+        
+        preencherAvaliacao(avaliacao);
     }
+
 
     /**
      * This method is called from within the constructor to initialize the form.
@@ -58,12 +61,12 @@ public class CadastrarProva extends javax.swing.JFrame {
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
-        popUpMenu = new javax.swing.JPopupMenu();
-        mnuEditar = new javax.swing.JMenuItem();
-        mnuExcluir = new javax.swing.JMenuItem();
-        pnlCadastroProva = new javax.swing.JPanel();
-        pnlPesquisaAluno = new javax.swing.JPanel();
-        pnlProva = new javax.swing.JPanel();
+        jPanel1 = new javax.swing.JPanel();
+        lblDetalhesProva = new javax.swing.JLabel();
+        scrListaAlunos1 = new javax.swing.JScrollPane();
+        lstDiscentes = new javax.swing.JList<>();
+        scrListaAlunos2 = new javax.swing.JScrollPane();
+        lstDisciplinas = new javax.swing.JList<>();
         pnlDetalhesProva = new javax.swing.JPanel();
         lblID = new javax.swing.JLabel();
         txtID = new javax.swing.JTextField();
@@ -85,41 +88,26 @@ public class CadastrarProva extends javax.swing.JFrame {
         lblParteOral = new javax.swing.JLabel();
         scrAprovado = new javax.swing.JScrollPane();
         txpAprovado = new javax.swing.JTextPane();
-        scrListaProvas = new javax.swing.JScrollPane();
-        lstAvaliacoes = new javax.swing.JList<>();
-        lblListaAlunos = new javax.swing.JLabel();
-        scrListaAlunos1 = new javax.swing.JScrollPane();
-        lstDiscentes = new javax.swing.JList<>();
-        scrListaAlunos2 = new javax.swing.JScrollPane();
-        lstDisciplinas = new javax.swing.JList<>();
-        lblProvasAplicadas = new javax.swing.JLabel();
-        lblDetalhesProva = new javax.swing.JLabel();
-
-        mnuEditar.setText("Editar");
-        mnuEditar.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                mnuEditarActionPerformed(evt);
-            }
-        });
-        popUpMenu.add(mnuEditar);
-
-        mnuExcluir.setText("Excluir");
-        mnuExcluir.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                mnuExcluirActionPerformed(evt);
-            }
-        });
-        popUpMenu.add(mnuExcluir);
 
         setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
-        setTitle("Provas");
-        setResizable(false);
 
-        pnlCadastroProva.setBackground(new java.awt.Color(255, 255, 255));
+        jPanel1.setBackground(new java.awt.Color(255, 255, 255));
 
-        pnlPesquisaAluno.setBackground(new java.awt.Color(255, 255, 255));
+        lblDetalhesProva.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lblDetalhesProva.setForeground(new java.awt.Color(0, 0, 0));
+        lblDetalhesProva.setText("Detalhes da prova");
 
-        pnlProva.setBackground(new java.awt.Color(255, 255, 255));
+        lstDiscentes.setBackground(new java.awt.Color(255, 255, 255));
+        lstDiscentes.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lstDiscentes.setForeground(new java.awt.Color(0, 0, 0));
+        lstDiscentes.setModel(lstDiscentesModel);
+        scrListaAlunos1.setViewportView(lstDiscentes);
+
+        lstDisciplinas.setBackground(new java.awt.Color(255, 255, 255));
+        lstDisciplinas.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lstDisciplinas.setForeground(new java.awt.Color(0, 0, 0));
+        lstDisciplinas.setModel(lstAulasModel);
+        scrListaAlunos2.setViewportView(lstDisciplinas);
 
         pnlDetalhesProva.setBackground(new java.awt.Color(255, 255, 255));
         pnlDetalhesProva.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(112, 112, 112)));
@@ -275,8 +263,6 @@ public class CadastrarProva extends javax.swing.JFrame {
                 .addContainerGap(17, Short.MAX_VALUE))
         );
 
-        pnlNotasLayout.linkSize(javax.swing.SwingConstants.VERTICAL, new java.awt.Component[] {scrAprovado, txtParteOral});
-
         javax.swing.GroupLayout pnlDetalhesProvaLayout = new javax.swing.GroupLayout(pnlDetalhesProva);
         pnlDetalhesProva.setLayout(pnlDetalhesProvaLayout);
         pnlDetalhesProvaLayout.setHorizontalGroup(
@@ -335,122 +321,48 @@ public class CadastrarProva extends javax.swing.JFrame {
                 .addContainerGap(13, Short.MAX_VALUE))
         );
 
-        lstAvaliacoes.setBackground(new java.awt.Color(255, 255, 255));
-        lstAvaliacoes.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lstAvaliacoes.setForeground(new java.awt.Color(0, 0, 0));
-        lstAvaliacoes.setModel(lstAvaliacaoModel);
-        lstAvaliacoes.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseReleased(java.awt.event.MouseEvent evt) {
-                lstAvaliacoesMouseReleased(evt);
-            }
-        });
-        scrListaProvas.setViewportView(lstAvaliacoes);
-
-        javax.swing.GroupLayout pnlProvaLayout = new javax.swing.GroupLayout(pnlProva);
-        pnlProva.setLayout(pnlProvaLayout);
-        pnlProvaLayout.setHorizontalGroup(
-            pnlProvaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlProvaLayout.createSequentialGroup()
-                .addGap(26, 26, 26)
-                .addComponent(scrListaProvas, javax.swing.GroupLayout.PREFERRED_SIZE, 297, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(pnlDetalhesProva, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
+        jPanel1.setLayout(jPanel1Layout);
+        jPanel1Layout.setHorizontalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addGap(50, 50, 50)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(scrListaAlunos1, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(scrListaAlunos2, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(60, 60, 60)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(lblDetalhesProva)
+                    .addComponent(pnlDetalhesProva, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(34, Short.MAX_VALUE))
+        );
+        jPanel1Layout.setVerticalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addGap(28, 28, 28)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel1Layout.createSequentialGroup()
+                        .addComponent(scrListaAlunos1, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(82, 82, 82)
+                        .addComponent(scrListaAlunos2, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(jPanel1Layout.createSequentialGroup()
+                        .addComponent(lblDetalhesProva)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(pnlDetalhesProva, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        pnlProvaLayout.setVerticalGroup(
-            pnlProvaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlProvaLayout.createSequentialGroup()
-                .addGroup(pnlProvaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(pnlDetalhesProva, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(scrListaProvas, javax.swing.GroupLayout.PREFERRED_SIZE, 373, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(0, 6, Short.MAX_VALUE))
-        );
-
-        lblListaAlunos.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lblListaAlunos.setForeground(new java.awt.Color(0, 0, 0));
-        lblListaAlunos.setText("Lista de alunos");
-
-        lstDiscentes.setBackground(new java.awt.Color(255, 255, 255));
-        lstDiscentes.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lstDiscentes.setForeground(new java.awt.Color(0, 0, 0));
-        lstDiscentes.setModel(lstDiscentesModel);
-        scrListaAlunos1.setViewportView(lstDiscentes);
-
-        lstDisciplinas.setBackground(new java.awt.Color(255, 255, 255));
-        lstDisciplinas.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lstDisciplinas.setForeground(new java.awt.Color(0, 0, 0));
-        lstDisciplinas.setModel(lstDisciplinasModel);
-        scrListaAlunos2.setViewportView(lstDisciplinas);
-
-        lblProvasAplicadas.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lblProvasAplicadas.setForeground(new java.awt.Color(0, 0, 0));
-        lblProvasAplicadas.setText("Provas aplicadas");
-
-        lblDetalhesProva.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
-        lblDetalhesProva.setForeground(new java.awt.Color(0, 0, 0));
-        lblDetalhesProva.setText("Detalhes da prova");
-
-        javax.swing.GroupLayout pnlPesquisaAlunoLayout = new javax.swing.GroupLayout(pnlPesquisaAluno);
-        pnlPesquisaAluno.setLayout(pnlPesquisaAlunoLayout);
-        pnlPesquisaAlunoLayout.setHorizontalGroup(
-            pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlProva, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addGroup(pnlPesquisaAlunoLayout.createSequentialGroup()
-                .addGap(0, 21, Short.MAX_VALUE)
-                .addGroup(pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(pnlPesquisaAlunoLayout.createSequentialGroup()
-                        .addComponent(scrListaAlunos1, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(scrListaAlunos2, javax.swing.GroupLayout.PREFERRED_SIZE, 450, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(pnlPesquisaAlunoLayout.createSequentialGroup()
-                        .addGroup(pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(pnlPesquisaAlunoLayout.createSequentialGroup()
-                                .addGap(4, 4, 4)
-                                .addComponent(lblListaAlunos))
-                            .addComponent(lblProvasAplicadas))
-                        .addGap(217, 217, 217)
-                        .addComponent(lblDetalhesProva)))
-                .addGap(23, 26, Short.MAX_VALUE))
-        );
-        pnlPesquisaAlunoLayout.setVerticalGroup(
-            pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlPesquisaAlunoLayout.createSequentialGroup()
-                .addGap(16, 16, 16)
-                .addComponent(lblListaAlunos)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(scrListaAlunos1, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(scrListaAlunos2, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGroup(pnlPesquisaAlunoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lblDetalhesProva, javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(lblProvasAplicadas, javax.swing.GroupLayout.Alignment.TRAILING))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(pnlProva, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-        );
-
-        javax.swing.GroupLayout pnlCadastroProvaLayout = new javax.swing.GroupLayout(pnlCadastroProva);
-        pnlCadastroProva.setLayout(pnlCadastroProvaLayout);
-        pnlCadastroProvaLayout.setHorizontalGroup(
-            pnlCadastroProvaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlPesquisaAluno, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        pnlCadastroProvaLayout.setVerticalGroup(
-            pnlCadastroProvaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnlCadastroProvaLayout.createSequentialGroup()
-                .addComponent(pnlPesquisaAluno, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(33, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlCadastroProva, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addGroup(layout.createSequentialGroup()
+                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 40, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlCadastroProva, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
         );
 
         pack();
@@ -464,128 +376,63 @@ public class CadastrarProva extends javax.swing.JFrame {
         // TODO add your handling code here:
     }//GEN-LAST:event_txtHoraActionPerformed
 
+    private void btnSalvarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSalvarActionPerformed
+        avaliacaoEmEdicao.setNotaParteEscrita(Integer.parseInt(txtParteEscrita.getText()));
+        avaliacaoEmEdicao.setNotaParteOral(txtParteOral.getText());
+        avaliacaoEmEdicao.setComentario(txtComentario.getText());
+        avaliacaoEmEdicao.setIdDiscente(lstDiscentes.getSelectedValue().getId().intValue());
+        avaliacaoEmEdicao.setIdAula(lstDisciplinas.getSelectedValue().getId().intValue());
+                
+        new AvaliacaoDao().salvar(avaliacaoEmEdicao);
+        cadastroProva.atualizarModelo(avaliacaoEmEdicao);
+
+        dispose();
+        
+
+    }//GEN-LAST:event_btnSalvarActionPerformed
+
     private void btnCancelarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnCancelarActionPerformed
         dispose();
     }//GEN-LAST:event_btnCancelarActionPerformed
 
     private void txtParteEscritaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtParteEscritaActionPerformed
-        if(estaAprovadoEscrita(Integer.parseInt(txtParteEscrita.getText())) && 
-                estaAprovadoOral(txtParteOral.getText())){
-            txpAprovado.setText("APROVADO");
-        }
-        else {
-            txpAprovado.setText("REPROVADO");
-        }
+        
     }//GEN-LAST:event_txtParteEscritaActionPerformed
 
     private void txtParteOralActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtParteOralActionPerformed
-        if(estaAprovadoEscrita(Integer.parseInt(txtParteEscrita.getText())) && 
-                estaAprovadoOral(txtParteOral.getText())){
-            txpAprovado.setText("APROVADO");
-        }
-        else {
-            txpAprovado.setText("REPROVADO");
-        }
+        
     }//GEN-LAST:event_txtParteOralActionPerformed
 
-    private void btnSalvarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSalvarActionPerformed
-        Avaliacao avaliacao = new Avaliacao();
-        avaliacao.setNotaParteEscrita(Integer.parseInt(txtParteEscrita.getText()));
-        avaliacao.setNotaParteOral(txtParteOral.getText());
-        avaliacao.setComentario(txtComentario.getText());
-        avaliacao.setIdDiscente(lstDiscentes.getSelectedValue().getId().intValue());
-        avaliacao.setIdAula(lstDisciplinas.getSelectedValue().getId().intValue());
-                
-        Long id = new AvaliacaoDao().salvar(avaliacao);
-        
-        avaliacao.setId(id);
-        
-        lstAvaliacaoModel.addElement(avaliacao);
-        
-        limparCampos();
-        
-    }//GEN-LAST:event_btnSalvarActionPerformed
-
-    private void lstAvaliacoesMouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_lstAvaliacoesMouseReleased
-        if (evt.getButton() == MouseEvent.BUTTON3
-                && lstAvaliacoes.getModel().getSize() > 0) {
-            indiceAvaliacaoSelecionada = lstAvaliacoes.locationToIndex(evt.getPoint());
-            lstAvaliacoes.setSelectedIndex(indiceAvaliacaoSelecionada);
-            popUpMenu.show(lstAvaliacoes, evt.getX(), evt.getY());
-        }
-    }//GEN-LAST:event_lstAvaliacoesMouseReleased
-
-    private void mnuExcluirActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_mnuExcluirActionPerformed
-        new AvaliacaoDao().excluir(lstAvaliacaoModel.get(indiceAvaliacaoSelecionada));
-
-        // Exclui graficamente 
-        lstAvaliacaoModel.remove(lstAvaliacoes.getSelectedIndex());
-    }//GEN-LAST:event_mnuExcluirActionPerformed
-
-    private void mnuEditarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_mnuEditarActionPerformed
-        EditarProva editarProva = new 
-        EditarProva(lstAvaliacaoModel.get(indiceAvaliacaoSelecionada), this, true);
-        
-        editarProva.setLocationRelativeTo(this);
-        editarProva.setVisible(true);
-    }//GEN-LAST:event_mnuEditarActionPerformed
-
-    private void limparCampos() {
-        // "Descrição" vazia
-        txtParteEscrita.setText(null);
-        txtParteOral.setText(null);
-        txtComentario.setText(null);
-        
-        // Seleção da "Descrição" para nova digitação
-        txtParteEscrita.requestFocus();
-    }
     
-    private boolean estaAprovadoEscrita(int parteEscrita){
-        if(parteEscrita >= 98 && parteEscrita <= 100){
-                return true;
-        }
-        return false;
+    public void preencherAvaliacao(Avaliacao avaliacao){
+        txtParteEscrita.setText(Integer.toString(avaliacao.getNotaParteEscrita()));
+        txtParteOral.setText(avaliacao.getNotaParteOral());
+        txtComentario.setText(avaliacao.getComentario());
+        lstDiscentes.setSelectedIndex(avaliacao.getIdDiscente());
+        lstDisciplinas.setSelectedIndex(avaliacao.getIdAula());
     }
-    
-    private boolean estaAprovadoOral(String parteOral){
-        return ("D".equals(parteOral) 
-                || "YD".equals(parteOral)
-                || "T".equals(parteOral));
-    }
-    
-    void atualizarModelo(Avaliacao avaliacao) {
-        lstAvaliacaoModel.set(indiceAvaliacaoSelecionada, avaliacao);
-    }
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnCancelar;
     private javax.swing.JButton btnSalvar;
+    private javax.swing.JPanel jPanel1;
     private javax.swing.JLabel lblAula;
     private javax.swing.JLabel lblComentarios;
     private javax.swing.JLabel lblDataHora;
     private javax.swing.JLabel lblDetalhesProva;
     private javax.swing.JLabel lblID;
-    private javax.swing.JLabel lblListaAlunos;
     private javax.swing.JLabel lblNotas;
     private javax.swing.JLabel lblParteEscrita;
     private javax.swing.JLabel lblParteOral;
-    private javax.swing.JLabel lblProvasAplicadas;
-    private javax.swing.JList<Avaliacao> lstAvaliacoes;
     private javax.swing.JList<Discente> lstDiscentes;
-    private javax.swing.JList<Aula> lstDisciplinas;
-    private javax.swing.JMenuItem mnuEditar;
-    private javax.swing.JMenuItem mnuExcluir;
-    private javax.swing.JPanel pnlCadastroProva;
+    private javax.swing.JList<br.edu.ifnmg.poo.trabalhofinalpoo.entity.Aula> lstDisciplinas;
     private javax.swing.JPanel pnlDetalhesProva;
     private javax.swing.JPanel pnlNotas;
-    private javax.swing.JPanel pnlPesquisaAluno;
-    private javax.swing.JPanel pnlProva;
-    private javax.swing.JPopupMenu popUpMenu;
     private javax.swing.JScrollPane scrAprovado;
     private javax.swing.JScrollPane scrComentario;
     private javax.swing.JScrollPane scrListaAlunos1;
     private javax.swing.JScrollPane scrListaAlunos2;
-    private javax.swing.JScrollPane scrListaProvas;
     private javax.swing.JTextPane txpAprovado;
     private javax.swing.JTextField txtAula;
     private javax.swing.JTextPane txtComentario;
